Seperate logic out from the display components
